requirements.txt
absl-py==2.1.0
alabaster==0.7.13
ale-py==0.8.1
astor==0.8.1
Babel==2.14.0
catkin-pkg==1.0.0
causal-world==1.2
charset-normalizer==3.4.1
cloudpickle==2.2.1
colorama==0.4.6
cycler==0.11.0
docutils==0.19
exceptiongroup==1.2.2
fonttools==4.38.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.62.3
gym==0.21.0
gym-notices==0.0.8
h5py==3.8.0
idna==3.10
imagesize==1.4.1
importlib-metadata==3.10.1
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.6
joblib==1.3.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mpi4py==4.0.3
numpy==1.21.6
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pluggy==1.2.0
protobuf==3.20.3
psutil==7.0.0
pybullet==2.8.5
pygame==2.1.0
Pygments==2.17.2
pyparsing==3.1.4
PyPrind==2.11.3
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0
rlpyt_lib
scipy==1.7.3
six==1.17.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stable-baselines==2.10.2
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==2.3.0
tomli==2.0.1
torch==1.13.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
